Defer typechecking of calls till unification. Pretty-print data structures that miscompare during tests.
This way, we have a fighting chance of diagnosing the problem when, as is common, it's deep in a big, hairy, nested structure. To do the pretty printing, we bring in ppx_deriving and use its `show` facility. To get PPX working, we switch to dune; I had no luck otherwise. It simplifies the makefile as well. Perhaps even more of the makefile can be replaced. Had to remove many unused var names, because they made something (the PPX rewriters?) throw errors.